split off trajectory generators
